# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Seattle Solvers Tools API",
    "description": "FTC parts inventory, search, and receipt reading for tools.seattlesolvers.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/admin/stats": {
      "get": {
        "summary": "Usage stats — request volume, active teams, receipt throughput",
        "tags": [
          "admin"
        ],
        "description": "Backed by RequestLog, one row per API request (see src/server.ts's onResponse hook; /health is excluded). Meant for capacity planning as more teams come on, not per-second monitoring — query it occasionally, not in a tight poll loop (that would itself show up in next week's numbers).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "generatedAt": {
                      "type": "string"
                    },
                    "requests": {
                      "type": "object",
                      "properties": {
                        "today": {
                          "type": "integer"
                        },
                        "last7Days": {
                          "type": "integer"
                        },
                        "serverErrorsLast7Days": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "today",
                        "last7Days",
                        "serverErrorsLast7Days"
                      ],
                      "additionalProperties": false
                    },
                    "teams": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "activeLast7Days": {
                          "type": "integer"
                        },
                        "withAnthropicKeyConfigured": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "total",
                        "activeLast7Days",
                        "withAnthropicKeyConfigured"
                      ],
                      "additionalProperties": false
                    },
                    "receipts": {
                      "type": "object",
                      "properties": {
                        "last7Days": {
                          "type": "integer"
                        },
                        "totalAllTime": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "last7Days",
                        "totalAllTime"
                      ],
                      "additionalProperties": false
                    },
                    "topTeamsLast7Days": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "teamId": {
                            "type": "string"
                          },
                          "teamNumber": {
                            "type": "integer",
                            "nullable": true
                          },
                          "teamName": {
                            "type": "string",
                            "nullable": true
                          },
                          "requestCount": {
                            "type": "integer"
                          }
                        },
                        "required": [
                          "teamId",
                          "teamNumber",
                          "teamName",
                          "requestCount"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "generatedAt",
                    "requests",
                    "teams",
                    "receipts",
                    "topTeamsLast7Days"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/teams": {
      "get": {
        "summary": "Every team using the site — name, number, and member count",
        "tags": [
          "admin"
        ],
        "description": "Every team that has ever signed up, not just the ones active in the last 7 days (see /admin/stats's topTeamsLast7Days for that narrower, activity-ranked view). Ordered by team number.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "number": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      },
                      "memberCount": {
                        "type": "integer"
                      }
                    },
                    "required": [
                      "id",
                      "number",
                      "name",
                      "memberCount"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/teams/{id}": {
      "delete": {
        "summary": "Remove a team from the site entirely",
        "tags": [
          "admin"
        ],
        "description": "Destructive and immediate, not reversible. Its inventory, receipts, expense history, and notifications are all deleted along with it. Any part it submitted to the shared library stays (other teams already rely on it) but loses its \"submitted by\" attribution. Members are not deleted — same as DELETE /teams/members/:userId, their accounts are kept but detached from the team (and demoted off TEAM_ADMIN, since that role only means something on a team); they can join another team with an invite code. Every member loses access immediately, on their very next request.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "enum": [
                    "null"
                  ],
                  "nullable": true
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/submissions": {
      "get": {
        "summary": "List parts submitted for the global library, pending review",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED"
              ],
              "default": "PENDING"
            },
            "in": "query",
            "name": "status",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "sku": {
                        "type": "string",
                        "nullable": true
                      },
                      "description": {
                        "type": "string",
                        "nullable": true
                      },
                      "productUrl": {
                        "type": "string",
                        "nullable": true
                      },
                      "purchaseUrl": {
                        "type": "string",
                        "nullable": true
                      },
                      "imageUrl": {
                        "type": "string",
                        "nullable": true
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "manufacturer": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name"
                        ],
                        "additionalProperties": false
                      },
                      "category": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name"
                        ],
                        "additionalProperties": false
                      },
                      "submittedByTeam": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "number": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "number",
                          "name"
                        ],
                        "additionalProperties": false,
                        "nullable": true
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "sku",
                      "description",
                      "productUrl",
                      "purchaseUrl",
                      "imageUrl",
                      "createdAt",
                      "manufacturer",
                      "category",
                      "submittedByTeam"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/submissions/{id}/approve": {
      "post": {
        "summary": "Approve a submission — publishes it to the global library for all teams",
        "tags": [
          "admin"
        ],
        "description": "A shared-library part carries one canonical price every team sees, so this is where a wrong one gets fixed: pass `priceOverride` to correct what the submitting team listed. Either way the team is told the outcome via an in-app notification (GET /notifications), and a corrected price says what it was changed from and to — a price silently rewritten under them would quietly skew that team's expense totals.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "not": {}
                  },
                  {
                    "type": "object",
                    "properties": {
                      "priceOverride": {
                        "type": "number",
                        "minimum": 0
                      },
                      "note": {
                        "type": "string",
                        "maxLength": 500
                      }
                    },
                    "additionalProperties": false,
                    "nullable": true
                  }
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "sku": {
                      "type": "string",
                      "nullable": true
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "productUrl": {
                      "type": "string",
                      "nullable": true
                    },
                    "purchaseUrl": {
                      "type": "string",
                      "nullable": true
                    },
                    "imageUrl": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "manufacturer": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name"
                      ],
                      "additionalProperties": false
                    },
                    "category": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name"
                      ],
                      "additionalProperties": false
                    },
                    "submittedByTeam": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "number": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "number",
                        "name"
                      ],
                      "additionalProperties": false,
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "sku",
                    "description",
                    "productUrl",
                    "purchaseUrl",
                    "imageUrl",
                    "createdAt",
                    "manufacturer",
                    "category",
                    "submittedByTeam"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}